+++ packages/expo-updates/src/Updates.ts
@@ -79,6 +79,16 @@ export const manifest: Partial<Manifest> =
   (ExpoUpdates.manifestString ? JSON.parse(ExpoUpdates.manifestString) : ExpoUpdates.manifest) ??
   {};
 
+/**
+ * If `expo-updates` is enabled, this is a `Date` object representing the creation time of the update that's currently running (whether it was embedded or downloaded at runtime).
+ *
+ * In development mode, or any other environment in which `expo-updates` is disabled, this value is
+ * null.
+ */
+export const createdAt: Date | null = ExpoUpdates.commitTime
+  ? new Date(ExpoUpdates.commitTime)
+  : null;
+
 const isUsingDeveloperTool = !!(manifest as any).developer?.tool;
 const isUsingExpoDevelopmentClient = NativeModulesProxy.ExponentConstants?.appOwnership === 'expo';
 const manualUpdatesInstructions = isUsingExpoDevelopmentClient